# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: leads
    description: Lead capture operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /marketing/leads:
    post:
      operationId: submitLead
      tags: [leads]
      summary: Submit a lead
      description: Capture a lead from the marketing site waitlist form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Lead submitted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    LeadInput:
      type: object
      required:
        - fullName
        - email
        - companyName
        - roleType
        - startupIntent
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 255
        companyName:
          type: string
          maxLength: 120
        roleType:
          type: string
          enum:
            - startup
            - mentor
            - investor
            - business_user
            - other
        startupIntent:
          type: array
          items:
            type: string
          minItems: 1
        servicesOffered:
          type: array
          items:
            type: string
        servicesNeeded:
          type: array
          items:
            type: string
        message:
          type: ["string", "null"]
          maxLength: 1000
    LeadResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
    ErrorResponse:
      type: object
      required:
        - ok
        - error
      properties:
        ok:
          type: boolean
        error:
          type: string