# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeadInputRoleType } from './leadInputRoleType';

export interface LeadInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  fullName: string;
  /** @maxLength 255 */
  email: string;
  /** @maxLength 120 */
  companyName: string;
  roleType: LeadInputRoleType;
  /** @minItems 1 */
  startupIntent: string[];
  servicesOffered?: string[];
  servicesNeeded?: string[];
  /**
     * @maxLength 1000
     * @nullable
     */
  message?: string | null;
}